+++ aur_scanner_gui/dialogs/__init__.py
@@ -1,3 +1,7 @@
 """
 Dialogs package for Cachy Security Suite.
 """
+from aur_scanner_gui.dialogs.download_dialog import AurDownloadDialog
+from aur_scanner_gui.dialogs.polkit_dialog import PolkitSetupDialog
+
+__all__ = ["AurDownloadDialog", "PolkitSetupDialog"]